Fix monitoring interval refresh to work on team tab and overview
